feat(release): allow user to now have manually resizable viewport
Previously, the viewport was programmatically resized thanks to how the code was written.
Now, if you wish to have a static sized viewport by hand and a percentage based card setup,
you may do so.

Keep in mind that if you use too high a value for the cards, it WILL extend past the viewport and
break your ability to see a specific number of cards.

Closes #7